# EventService.Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EventService",
    "version": "v1"
  },
  "paths": {
    "/api/Events/my-events": {
      "get": {
        "tags": [
          "Events"
        ],
        "operationId": "GetMyEvents",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Events/{eventId}": {
      "get": {
        "tags": [
          "Events"
        ],
        "operationId": "GetEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Events": {
      "post": {
        "tags": [
          "Events"
        ],
        "operationId": "CreateEvent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Events"
        ],
        "operationId": "UpdateEvent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Events/{eventId}/register": {
      "put": {
        "tags": [
          "Events"
        ],
        "operationId": "Register",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Events/{eventId}/deregister": {
      "put": {
        "tags": [
          "Events"
        ],
        "operationId": "DeRegister",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Events/{eventId}/Users/{userId}": {
      "delete": {
        "tags": [
          "Events"
        ],
        "operationId": "KickUser",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Events/{eventId}/waiting-list": {
      "get": {
        "tags": [
          "Events"
        ],
        "operationId": "GetWaitingList",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventUserReadDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventUserReadDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventUserReadDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Events/{eventId}/approved-list": {
      "get": {
        "tags": [
          "Events"
        ],
        "operationId": "GetApprovedList",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventUserReadDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventUserReadDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventUserReadDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Events/update-event-user": {
      "post": {
        "tags": [
          "Events"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUserUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUserUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventUserUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Events/nearby-events": {
      "post": {
        "tags": [
          "Events"
        ],
        "operationId": "GetNearbyEvents",
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLocationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLocationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventReadDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventReadDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventReadDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Events/events-in-area": {
      "post": {
        "tags": [
          "Events"
        ],
        "operationId": "GetEventsInArea",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AreaDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AreaDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AreaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventReadDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventReadDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventReadDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Events/types": {
      "get": {
        "tags": [
          "Events"
        ],
        "operationId": "GetEventTypes",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventTypeDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventTypeDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventTypeDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EventTypeDto": {
        "type": "object",
        "additionalProperties": false
      },
      "EventUserReadDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventReadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "eventType": {
            "$ref": "#/components/schemas/EventTypeDto"
          },
          "eventUsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventUserReadDto"
            },
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "buddiesCount": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "EventCreateDto": {
        "required": [
          "eventType",
          "latitude",
          "longitude",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "eventType": {
            "$ref": "#/components/schemas/EventTypeDto"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "EventUpdateDto": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "eventType": {
            "$ref": "#/components/schemas/EventTypeDto"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "EventUserUpdateDto": {
        "required": [
          "approved",
          "eventId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "approved": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UserLocationDto": {
        "required": [
          "latitude",
          "longitude"
        ],
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "AreaDto": {
        "required": [
          "northEastLocation",
          "southWestLocation"
        ],
        "type": "object",
        "properties": {
          "northEastLocation": {
            "$ref": "#/components/schemas/UserLocationDto"
          },
          "southWestLocation": {
            "$ref": "#/components/schemas/UserLocationDto"
          }
        },
        "additionalProperties": false
      }
    }
  }
}